Move shorttmpdir fixture into conftest.py
